Microphone device filtering heuristic is too harsh https://bugs.webkit.org/show_bug.cgi?id=234596 <rdar://problem/86803287>
Reviewed by Eric Carlson.

Source/WebCore:

When audio capture happens, there are more audio devices and some audio devices get an input stream for echo cancellation.
We add an API to disable that filtering which is used when capture is not done in the process where audio enumeration is done.
A longer term solution should look at doing this filtering in all cases.

Manually tested in Safari and Minibrowser.

* WebCore.xcodeproj/project.pbxproj:
* platform/mediastream/mac/CoreAudioCaptureDeviceManager.cpp:
* platform/mediastream/mac/CoreAudioCaptureDeviceManager.h:

Source/WebKit:

* UIProcess/WebPageProxy.cpp:
Disable filtering if audio capture happens in GPUProcess.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287715 268f45cc-cd09-0410-ab3c-d52691b4dbfc