Separate "linked-on-or-{before, after}-everything" override from the SDK version https://bugs.webkit.org/show_bug.cgi?id=234930
Reviewed by Wenson Hsieh.

Source/WebCore:

* platform/cocoa/VersionChecks.mm:
(WebCore::linkedOnOrAfter):
Make use of the new bit instead of checking for Safari's bundle identifier.

* testing/js/WebCoreTestSupport.cpp:
(WebCoreTestSupport::setLinkedOnOrAfterEverythingForTesting):
Adopt the new bit.

Source/WebKit:

* Shared/AuxiliaryProcess.h:
* Shared/Cocoa/AuxiliaryProcessCocoa.mm:
(WebKit::AuxiliaryProcess::platformInitialize):
* Shared/EntryPointUtilities/Cocoa/XPCService/XPCServiceEntryPoint.h:
(WebKit::XPCServiceInitializer):
* Shared/EntryPointUtilities/Cocoa/XPCService/XPCServiceEntryPoint.mm:
(WebKit::XPCServiceInitializerDelegate::getLinkedOnOrAfterOverride):
Plumb the LOA override optional through process initialization (just like
we do for SDK version, because linkedOnOrAfter() is used VERY early in
JavaScriptCore initialization).

* UIProcess/API/Cocoa/WKProcessPool.mm:
(+[WKProcessPool _setLinkedOnOrBeforeEverythingForTesting]):
(+[WKProcessPool _setLinkedOnOrAfterEverythingForTesting]):
(+[WKProcessPool _setLinkedOnOrAfterEverything]):
* UIProcess/API/Cocoa/WKProcessPoolPrivate.h:
Expose non-testing SPI for Safari to opt in to "linked-on-or-after-everything"
so that we can (eventually) stop using a bundle identifier check.

* UIProcess/Launcher/cocoa/ProcessLauncherCocoa.mm:
(WebKit::ProcessLauncher::launchProcess):

* UIProcess/WebProcessPool.cpp:
Initialize the LOA override using the same logic we previously used in linkedOnOrAfter().

Source/WTF:

* wtf/cocoa/RuntimeApplicationChecksCocoa.cpp:
(WTF::linkedOnOrAfterOverrideValue):
(WTF::setLinkedOnOrAfterOverride):
(WTF::linkedOnOrAfterOverride):
* wtf/cocoa/RuntimeApplicationChecksCocoa.h:
Add an explicit optional bit for "this application should pretend to be
linked on or {after, before} everything" instead of just overriding the
SDK version. This is a prerequisite for getting rid of applicationSDKVersion(),
which will happen in a future patch.

This is in WTF because linkedOnOrAfter() is going to move here in a
future patch as well.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287720 268f45cc-cd09-0410-ab3c-d52691b4dbfc