REGRESSION (Safari 14): background-attachment: local does not work https://bugs.webkit.org/show_bug.cgi?id=219324 Source/WebCore:
<rdar://71808922>

Reviewed by Antti Koivisto.

With async overflow scrolling, we need to trigger a repaint if the scrolling element
has any background layer with `background-attachment: local`.

The background won't always be synchronized with the scrolling, but a little jitter
is better than a broken CSS behavior (and this is what Firefox does).

Test: fast/repaint/background-attachment-local-scroll.html

* rendering/RenderLayerBacking.cpp:
(WebCore::RenderLayerBacking::needsRepaintOnCompositedScroll const):
* rendering/style/RenderStyle.cpp:
(WebCore::RenderStyle::hasAnyLocalBackground const):
* rendering/style/RenderStyle.h:

LayoutTests:

Reviewed by Antti Koivisto.

Skip the test on platforms without async overflow scroll.

* TestExpectations:
* fast/repaint/background-attachment-local-scroll-expected.txt: Added.
* fast/repaint/background-attachment-local-scroll.html: Added.
* platform/ios-wk2/TestExpectations:
* platform/mac-wk2/TestExpectations:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287757 268f45cc-cd09-0410-ab3c-d52691b4dbfc