Expose rvfc processingDuration for the WebRTC code path https://bugs.webkit.org/show_bug.cgi?id=234949
Reviewed by Eric Carlson.

Source/WebCore:

This processing duration is the time it took for an assembled frame to be decoded and sent to RealtimIncomingVideoSource.
This duration can be approximately retrieved from WebRTC encoded transform (time for the assembled frame, aka time just before decoding)
and the time at which it is exposed by requestVideoFrameCallback (aka time just after decoding).
The processing duration is rounded in milliseconds.

Covered by updated test.

* platform/mediastream/RealtimeIncomingVideoSource.cpp:

LayoutTests:

* webrtc/peerConnection-rvfc.html:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287759 268f45cc-cd09-0410-ab3c-d52691b4dbfc