[LFC][IFC] Implement TextUtil::breakWord for the complex font codepath using ubrk_next https://bugs.webkit.org/show_bug.cgi?id=234998
Reviewed by Antti Koivisto.

In order to use ubrk_preceding/ubrk_following with arbitrary position (binary search) inside a cluster with surrogate pairs we have to
implement some additional surrogate boundary checks (it would move the index to the beginning/end of the surrogate when it falls right
in the middle of it). In addition to that, ICU would still need to scan the content from the start to find the right index for
the boundary.
This new breakWord implementation simply iterates over the clusters by calling ubrk_next until the content overflows.

* layout/formattingContexts/inline/text/TextUtil.cpp:
(WebCore::Layout::TextUtil::breakWord):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287809 268f45cc-cd09-0410-ab3c-d52691b4dbfc