https://brianpeiris.github.io/spatial-audio-test/?ramped stops playing after a few seconds https://bugs.webkit.org/show_bug.cgi?id=234979
Reviewed by Eric Carlson.

Source/WebCore:

In ConeEffect::gain(), due to precision issues and with certain panner node parameters, it was possible for
`sourceToListener.dot(sourceOrientation)` to return a value that is very slightly outside the [-1.0, 1.0]
range. We would then call `acos()` on the dot product and it would thus return NaN.

To addresss this, we now make sure the clamp the dot product to the expected range before calling acos().

I also made the following stylistic changes:
- Drop unnecessary normalizedSourceOrientation local variable since the sourceOrientation paramter is not const
- Add missing curly brackets in if conditions with more than one lines (due to comments)
- Call rad2deg() for readability instead of duplicating its logic here

Test: webaudio/Panner/panner-cone-gain-nan.html

* platform/audio/Cone.cpp:
(WebCore::ConeEffect::gain const):

LayoutTests:

Add layout test coverage.

* webaudio/Panner/panner-cone-gain-nan-expected.txt: Added.
* webaudio/Panner/panner-cone-gain-nan.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287854 268f45cc-cd09-0410-ab3c-d52691b4dbfc