Followup to r287863 - adjust line wrapping behavior in image overlays https://bugs.webkit.org/show_bug.cgi?id=235035 rdar://85139146
Reviewed by Tim Horton.

Source/WebCore:

The previous change I landed in r287863 was written under the assumption that `-shouldWrap` describes whether or
not a recognized line of text should wrap to the next line. However, after trying it out, it seems to indicate
whether or not the previous line of text should wrap to the current line. As a result, the current adoption of
this property is wrong, since the line wrapping is off-by-one.

Fix this by renaming the `shouldWrap` boolean flag in TextRecognitionLineData to the much less ambiguous name
`hasTrailingNewline`, and set `hasTrailingNewline` based on the `shouldWrap` property of the *next*
VKWKLineInfo.

* dom/ImageOverlay.cpp:
(WebCore::ImageOverlay::updateSubtree):

Avoid a potentially even-more-confusing `!= !!` here by using `static_cast<>` to check the nullity of the
`RefPtr`.

* platform/TextRecognitionResult.h:
(WebCore::TextRecognitionLineData::TextRecognitionLineData):
(WebCore::TextRecognitionLineData::encode const):
(WebCore::TextRecognitionLineData::decode):

Make sure we also flip the default from `false` to `true`, since `hasTrailingNewline` is intended to have the
opposite effect as the extant `shouldWrap`.

* testing/Internals.cpp:
(WebCore::makeDataForLine):
* testing/Internals.h:
* testing/Internals.idl:

Source/WebKit:

Use the `-shouldWrap` property to compute `hasTrailingNewline`. See WebCore/ChangeLog for more details.

* Platform/cocoa/TextRecognitionUtilities.mm:
(WebKit::makeTextRecognitionResult):

LayoutTests:

Make a slight adjustment to an existing layout test (i.e. rename `shouldWrap` to `hasTrailingNewline`). See
WebCore/ChangeLog for more details.

* fast/images/text-recognition/image-overlay-line-wrapping.html:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287866 268f45cc-cd09-0410-ab3c-d52691b4dbfc