_paintBehaviorForDestinationContext should consider kCGContextTypeDisplayList type contexts as potentially being snapshots https://bugs.webkit.org/show_bug.cgi?id=234626 <rdar://83501315>
Patch by Matt Woodrow <mattwoodrow@apple.com> on 2022-01-12
Reviewed by Simon Fraser.

_paintBehaviorForDestinationContext currently always returns `Normal` for contexts that aren't a bitmap context.
AppKit now provides a context that is of type kCGContextTypeDisplayList, and we want to still potentially consider it a
snapshot paint (if the other conditions below are met). This ensure images are sync decoded for consumers that just
want a single paint.

* WebView/WebFrame.mm:
(-[WebFrame _paintBehaviorForDestinationContext:]):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287951 268f45cc-cd09-0410-ab3c-d52691b4dbfc