[css-grid] Fix rounding of distributed free space to flexible tracks https://bugs.webkit.org/show_bug.cgi?id=234917
LayoutTests/imported/w3c:

Reviewed by Darin Adler.

Source/WebCore:

Reviewed by Darin Adler.

When computing the growth size for flex sized tracks, the flexFraction multiplied by the flex factor can result
in a non-integer size. However, we floor the stretched size to fit in a LayoutUnit. This means that we may lose
the fractional part of the computation which can cause the entire free space not being distributed evenly. This
fix is to sum up the leftover fractional part from every flexible track to avoid this issue.

It is an import of Chromium GridNG CL at  https://chromium-review.googlesource.com/c/chromium/src/+/3193674.

* rendering/GridTrackSizingAlgorithm.cpp:
(WebCore::GridTrackSizingAlgorithm::computeFlexSizedTracksGrowth const):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287977 268f45cc-cd09-0410-ab3c-d52691b4dbfc