[GTK][WPE][WebGL2] compilation fixes https://bugs.webkit.org/show_bug.cgi?id=225563
Patch by Alejandro G. Castro <alex@igalia.com> on 2022-01-13
Reviewed by Chris Lord.

Fix the compilation and add the ANGLE dependency, the architecture
is still not the final one but it is handy to use WebGL2 demos for
testing already.

.:

* Source/cmake/OptionsGTK.cmake: Added dependency with ANGLE.
* Source/cmake/OptionsWPE.cmake: Added dependency with
ANGLE. Remove the default ENABLE_WEBGL2 because it is the same
default value in WebKitFeatures.cmake.

Source/WebCore:

We are still not enabling the tests until we have the final
architecture.

* html/canvas/WebGLRenderingContextBase.cpp:
(WebCore::WebGLRenderingContextBase::drawElements): Leftover code,
there is no port using OpenGL directly to support WebGL2 anymore.
(WebCore::WebGLRenderingContextBase::drawElementsInstanced):
Ditto.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287983 268f45cc-cd09-0410-ab3c-d52691b4dbfc