html/semantics/scripting-1/the-script-element/script-type-and-language-empty.html WPT test is failing https://bugs.webkit.org/show_bug.cgi?id=235202
Reviewed by Darin Adler.

LayoutTests/imported/w3c:

Rebaseline test that is now passing.

* web-platform-tests/html/semantics/scripting-1/the-script-element/script-type-and-language-empty-expected.txt:

Source/WebCore:

Per https://html.spec.whatwg.org/multipage/scripting.html#prepare-a-script (step 8), we should assume
the type is "text/javascript" if the type attribute is present and its value is the empty string.
Our logic in ScriptElement::determineScriptType() was slightly wrong and didn't get this case right
because it was doing an `isEmpty()` check which could mean both that the attribute is missing or present
with a value that is the empty string.

No new tests, rebaselined existing test.

* dom/ScriptElement.cpp:
(WebCore::ScriptElement::determineScriptType const):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@288004 268f45cc-cd09-0410-ab3c-d52691b4dbfc