Stop forcing the creation of PseudoElement in KeyframeEffect https://bugs.webkit.org/show_bug.cgi?id=235225
Reviewed by Antti Koivisto.

At the time we started supporting targeting pseudo-elements using the Web Animations API,
we resorted to some hack-ish code to enforce the creation of a PseudoElement should the
target of the animation be a ::before or ::after pseudo-element.

This isn't necessary, and was due to some incorrect code in
RenderTreeUpdater::GeneratedContent::updatePseudoElement() which this patch corrects.
Indeed, if there are animations targeting the pseudo-element in question, we should not
be removing the pseudo-element and exiting this method.

* animation/KeyframeEffect.cpp:
(WebCore::KeyframeEffect::didChangeTargetStyleable):
(WebCore::elementOrPseudoElementForStyleable): Deleted.
* rendering/updating/RenderTreeUpdaterGeneratedContent.cpp:
(WebCore::RenderTreeUpdater::GeneratedContent::updatePseudoElement):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@288016 268f45cc-cd09-0410-ab3c-d52691b4dbfc