[GStreamer] tests media/track/audio-track-configuration.html and media/track/video-track-configuration.html fail https://bugs.webkit.org/show_bug.cgi?id=234084
Patch by Philippe Normand <pnormand@igalia.com> on 2022-01-17
Reviewed by Xabier Rodriguez-Calvar.

Source/WebCore:

The known caps and tags of the track GstStream are now exposed through the corresponding
Platform{Audio,Video}TrackConfiguration. This is supported only when playbin3 is used at
runtime. Supporting playbin2 would be some work (pad probes upfront of decoders?) that could
be done as a follow-up. The codec string is generated with the new GStreamer API
gst_codec_utils_caps_get_mime_codec(), which will ship in 1.20.

* platform/graphics/gstreamer/AudioTrackPrivateGStreamer.cpp:
(WebCore::AudioTrackPrivateGStreamer::AudioTrackPrivateGStreamer):
* platform/graphics/gstreamer/VideoTrackPrivateGStreamer.cpp:
(WebCore::VideoTrackPrivateGStreamer::VideoTrackPrivateGStreamer):

LayoutTests:

* media/track/audio-track-configuration.html: Add a timeout to each promise, allowing for
nicer actual results when a sub-set of the promises time out. Without this a single promise
timeout would make the entire test time out and the actual results would not be reported
properly.
* media/track/video-track-configuration.html: Ditto.
* platform/glib/TestExpectations: Update {audio,video}-track expectations, tests still
expected to fail but some improvements will be available when we switch to playbin3.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@288092 268f45cc-cd09-0410-ab3c-d52691b4dbfc