AX: Expose toggle buttons using role="button" as form controls. https://bugs.webkit.org/show_bug.cgi?id=234119
Patch by Sepand Parhami <sparhami@apple.com> on 2022-01-17
Reviewed by Chris Fleizach.

Source/WebCore:

Add the ToggleButton role to the list of aria controls in isARIAControl so that it
is a control when checked by the accessibility object search.

Tests: LayoutTests/accessibility/mac/search-predicate-element-count.html

* accessibility/AccessibilityNodeObject.cpp:
(WebCore::AccessibilityNodeObject::isLabelable const):

Remove redundant check, isControl already covers it.

* accessibility/AccessibilityObject.cpp:
(WebCore::AccessibilityObject::isARIAControl):

LayoutTests:

Added a test case for button/role="button" with aria-presssed. Reordered a few things as
the previously named "inputs" section was not what was being checked in the test (it was
the number of buttons) and it had previously matched the number by coincidence.

* accessibility/mac/search-predicate-element-count-expected.txt:
* accessibility/mac/search-predicate-element-count.html:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@288100 268f45cc-cd09-0410-ab3c-d52691b4dbfc